@@ -7,7 +7,7 @@ import 'model_carousel_item_animation.dart';
 class CarouselSlide1 extends StatefulWidget {
   final int slideDuration;
 
-  const CarouselSlide1({Key key, @required this.slideDuration})
+  const CarouselSlide1({Key? key, required this.slideDuration})
       : super(key: key);
 
   @override
@@ -16,8 +16,8 @@ class CarouselSlide1 extends StatefulWidget {
 
 class _CarouselSlide1State extends State<CarouselSlide1>
     with SingleTickerProviderStateMixin {
-  Animation<double> animation;
-  AnimationController animationController;
+  late Animation<double> animation;
+  late AnimationController animationController;
 
   List<SlideItemAnimationModel> slideItems = [
     SlideItemAnimationModel(
